Push for reinstallation of windows

+++ src/commands/moderation/warn.js
@@ -1,6 +1,5 @@
-const { SlashCommandBuilder } = require('discord.js');
+const { SlashCommandBuilder, ActionRowBuilder, ButtonBuilder, ButtonStyle } = require('discord.js');
 const { Infraction } = require('../../structures/schemas/infraction');
-const { moderation: modButtons } = require('../../util/builders/components');
 const { errorEmbed, warningEmbed } = require('../../util/builders/embeds');
 const { successEmbed, moderation: modEmbeds } = require('../../util/builders/embeds');
 const { channels, roles } = require('../../config.json');
@@ -14,7 +13,7 @@ const info = new SlashCommandBuilder()
     .addStringOption((option) => option.setName('warning').setDescription('The reason you are warning this user').setRequired(true));
 
 const execute = async (inter) => {
-    const { options, user: modUser, guild } = inter;
+    const { options, user: modUser, guild, channel } = inter;
     const targetMember = inter.options.getMember('user');
     const { user: targetUser } = targetMember;
 
@@ -43,14 +42,33 @@ const execute = async (inter) => {
 
     const warningAmount = await Infraction.count({ target: targetUser.id });
     if (warningAmount >= 3) {
-        // TODO: Respond DM with ban infraction entry
-        // await targetUser.send(ban embed here);
-        await targetMember.ban({ deleteMessageSeconds: 604800, reason: `[ScrappyBot] User reached ${warningAmount} warnings` });
-        autoUnban(guild, targetUser.id, 2592000000);
+        const buttonNumber = generateId(5, true);
+        const autobanButtons = new ActionRowBuilder().addComponents(
+            new ButtonBuilder().setCustomId(`confirmBan_${buttonNumber}`).setLabel('Confirm').setStyle(ButtonStyle.Success),
+            new ButtonBuilder().setCustomId(`denyBan_${buttonNumber}`).setLabel('Deny').setStyle(ButtonStyle.Danger)
+        );
+        const autobanMsg = await inter.followUp({
+            embeds: [warningEmbed(`User reached ${warningAmount} warnings. User will be banned automatically in 30 seconds if no choice has been made. Ban user?`)],
+            components: [autobanButtons],
+        });
 
-        // TODO: add confirm button for ban
-        await inter.followUp({ embeds: [warningEmbed(`User was automatically banned for reaching ${warningAmount} warnings`)] });
-        await Infraction.deleteMany({ target: targetUser.id });
+        const interFilter = (i) => i.customId === `confirmBan_${buttonNumber}` || i.customId === `denyBan_${buttonNumber}`;
+        const collector = channel.createMessageComponentCollector({ interFilter, time: 30000 });
+        collector.on('collect', async (i) => {
+            if (i.user.id !== modUser.id) return i.reply({ content: 'That button is not for you!', ephemeral: true });
+            if (i.customId === `confirmBan_${buttonNumber}`) {
+                // TODO: Respond DM with ban infraction entry
+                // await targetUser.send(ban embed here);
+
+                targetMember.ban({ deleteMessageSeconds: 604800, reason: `[ScrappyBot] User reached ${warningAmount} warnings` });
+                autoUnban(guild, targetUser.id, 2592000000);
+
+                autobanMsg.edit({ embeds: [successEmbed(`**${targetUser.tag}** was automatically banned for reaching ${warningAmount} warnings`)], components: [] });
+                await Infraction.deleteMany({ target: targetUser.id });
+            } else if (i.customId === `denyBan_${buttonNumber}`) {
+                autobanMsg.edit({ embeds: [errorEmbed('User has not been banned')], components: [] });
+            }
+        });
     }
     if (warningAmount < 3 && warningAmount !== 0) {
         targetMember.roles.add(roles.warnings[warningAmount]);
